+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/ic.kt
@@ -8,6 +8,7 @@ import org.jetbrains.kotlin.incremental.components.LocationInfo
 import org.jetbrains.kotlin.incremental.components.LookupTracker
 import org.jetbrains.kotlin.incremental.components.Position
 import org.jetbrains.kotlin.incremental.components.ScopeKind
+import org.jetbrains.kotlin.ir.declarations.IrClass
 import org.jetbrains.kotlin.ir.declarations.IrDeclaration
 import org.jetbrains.kotlin.ir.declarations.IrFunction
 import org.jetbrains.kotlin.ir.declarations.IrSimpleFunction
@@ -15,6 +16,27 @@ import org.jetbrains.kotlin.ir.util.kotlinFqName
 import org.jetbrains.kotlin.ir.util.parentAsClass
 import org.jetbrains.kotlin.name.FqName
 import org.jetbrains.kotlin.psi.doNotAnalyze
+
+/**
+ * Tracks a call from one [callingDeclaration] to a [calleeClass] to inform incremental compilation.
+ */
+internal fun IrMetroContext.trackClassLookup(
+  callingDeclaration: IrDeclaration,
+  calleeClass: IrClass,
+) {
+  callingDeclaration.withAnalyzableKtFile { filePath ->
+    trackLookup(
+      container = calleeClass.parent.kotlinFqName,
+      declarationName = calleeClass.name.asString(),
+      scopeKind = ScopeKind.PACKAGE,
+      location =
+        object : LocationInfo {
+          override val filePath = filePath
+          override val position: Position = Position.NO_POSITION
+        },
+    )
+  }
+}
 
 /**
  * Tracks a call from one [callingDeclaration] to a [calleeFunction] to inform incremental
@@ -27,16 +49,13 @@ internal fun IrMetroContext.trackFunctionCall(
   callingDeclaration: IrDeclaration,
   calleeFunction: IrFunction,
 ) {
-  val ktFile = callingDeclaration.fileParentOrNull?.getKtFile()
-  if ((ktFile != null && ktFile.doNotAnalyze == null) || ktFile == null) {
-    // Not every declaration has a file parent, for example IR-generated accessors
-    val filePath =
-      ktFile?.virtualFile?.path ?: callingDeclaration.fileParentOrNull?.fileEntry?.name ?: return
+  callingDeclaration.withAnalyzableKtFile { filePath ->
     val declaration =
       (calleeFunction as? IrSimpleFunction)?.correspondingPropertySymbol?.owner ?: calleeFunction
     trackLookup(
       container = calleeFunction.parentAsClass.kotlinFqName,
-      functionName = declaration.name.asString(),
+      declarationName = declaration.name.asString(),
+      scopeKind = ScopeKind.CLASSIFIER,
       location =
         object : LocationInfo {
           override val filePath = filePath
@@ -48,7 +67,8 @@ internal fun IrMetroContext.trackFunctionCall(
 
 internal fun IrMetroContext.trackLookup(
   container: FqName,
-  functionName: String,
+  declarationName: String,
+  scopeKind: ScopeKind,
   location: LocationInfo,
 ) {
   withLookupTracker {
@@ -56,12 +76,23 @@ internal fun IrMetroContext.trackLookup(
       filePath = location.filePath,
       position = if (requiresPosition) location.position else Position.NO_POSITION,
       scopeFqName = container.asString(),
-      scopeKind = ScopeKind.CLASSIFIER,
-      name = functionName,
+      scopeKind = scopeKind,
+      name = declarationName,
     )
   }
 }
 
 internal inline fun IrMetroContext.withLookupTracker(body: LookupTracker.() -> Unit) {
   lookupTracker?.let { tracker -> synchronized(tracker) { tracker.body() } }
+}
+
+private fun IrDeclaration.withAnalyzableKtFile(body: (filePath: String) -> Unit) {
+  val callingDeclaration = this
+  val ktFile = callingDeclaration.fileParentOrNull?.getKtFile()
+  if ((ktFile != null && ktFile.doNotAnalyze == null) || ktFile == null) {
+    // Not every declaration has a file parent, for example IR-generated accessors
+    val filePath =
+      ktFile?.virtualFile?.path ?: callingDeclaration.fileParentOrNull?.fileEntry?.name ?: return
+    body(filePath)
+  }
 }